Updated structure names
