--- conf/wdio-on-prem.conf.js
@@ -1,10 +1,0 @@
-var defaults = require("./wdio.conf.js");
-var _ = require("lodash");
-
-var overrides = {
-  specs: [
-    './test/specs/plain/*.js'
-  ],
-};
-
-exports.config = _.defaultsDeep(overrides, defaults.config);
