+++ resources/conf/wdio-browserstack-parallel.conf.js
@@ -2,23 +2,45 @@ var defaults = require("./wdio.conf.js");
 var _ = require("lodash");
 
 var overrides = {
-  browserstackLocal: false,
   user: process.env.BROWSERSTACK_USERNAME || 'BROWSERSTACK_USERNAME',
   key: process.env.BROWSERSTACK_ACCESS_KEY || 'BROWSERSTACK_ACCESS_KEY',
   specs: [
     './test/specs/plain/*.js'
   ],
-  capabilities: [{
-    maxInstances: 1,
+  maxInstances: 5,
+  commonCapabilities: {
     'browserstack.debug': true,
     'browserstack.video': true,
+    acceptInsecureCerts: true
+  },
+  capabilities: [{
     os: "OS X",
     os_version: "Catalina",
     browserName: 'Chrome',
     browser_version: "latest",
-    acceptInsecureCerts: true,
-    name: 'BStack-Test',
-    build: 'BStack Build webdriverio single'
+    name: 'parallel_test',
+    build: 'webdriverio-browserstack',
+  },{
+    device: "Samsung Galaxy S20",
+    os_version: "10.0",
+    real_mobile: "true",
+    browserName: 'Android',
+    name: 'parallel_test',
+    build: 'webdriverio-browserstack',
+  },{
+    os: "Windows",
+    os_version: "10",
+    browserName: 'Chrome',
+    browser_version: "latest",
+    name: 'parallel_test',
+    build: 'webdriverio-browserstack',
+  },{
+    device: "iPhone 12",
+    os_version: "14",
+    real_mobile: "true",
+    browserName: 'iPhone',
+    name: 'parallel_test',
+    build: 'webdriverio-browserstack',
   }],
   afterTest: function (test, context, { error, result, duration, passed, retries }) {
     if(passed) {
