+++ resources/conf/wdio-docker-parallel.conf.js
